+++ docs/data/material/getting-started/templates/sign-in/theme/AppTheme.js
@@ -3,8 +3,8 @@
  * Generated by a script.
  */
 import * as React from 'react';
-import type {} from '@mui/material/themeCssVarsAugmentation';
-import { ThemeProvider, createTheme, alpha, Shadows } from '@mui/material/styles';
+
+import { ThemeProvider, createTheme, alpha } from '@mui/material/styles';
 import {
   inputsCustomizations,
   dataDisplayCustomizations,
@@ -15,35 +15,9 @@ import {
 // there will be no files imported here, so that the theme becomes a single file
 // and can be easily copied between projects
 
-declare module '@mui/material/Paper' {
-  interface PaperPropsVariantOverrides {
-    highlighted: true;
-  }
-}
-declare module '@mui/material/styles/createPalette' {
-  interface ColorRange {
-    50: string;
-    100: string;
-    200: string;
-    300: string;
-    400: string;
-    500: string;
-    600: string;
-    700: string;
-    800: string;
-    900: string;
-  }
-
-  interface PaletteColor extends ColorRange {}
-
-  interface TypeText {
-    warning: string;
-  }
-}
-
 const defaultTheme = createTheme();
 
-const customShadows: Shadows = [...defaultTheme.shadows];
+const customShadows = [...defaultTheme.shadows];
 
 export const brand = {
   50: 'hsl(210, 100%, 95%)',
@@ -274,13 +248,7 @@ const theme = createTheme({
   },
 });
 
-export default function AppTheme({
-  children,
-  disableCustomTheme,
-}: {
-  children: React.ReactNode;
-  disableCustomTheme?: boolean;
-}) {
+export default function AppTheme({ children, disableCustomTheme }) {
   if (disableCustomTheme) {
     return children;
   }
