+++ docs/pages/material-ui/getting-started/templates/sign-in.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import AppTheme from 'docs/src/modules/components/AppTheme';
 import TemplateFrame from 'docs/src/modules/components/TemplateFrame';
-import SignIn from 'docs/data/material/getting-started/templates/sign-in/SignIn';
+import SignIn from 'docs/data/material/getting-started/templates/sign-in/App';
 
 export default function Page() {
   return (
